requirements.txt
Flask==2.0.1
python-dotenv==0.19.2
requests_html==0.10.0
gunicorn==20.1.0